--- src/DistIL/IR/DSL/IInstructionMatcher.cs
@@ -1,6 +1,0 @@
-﻿namespace DistIL.IR.DSL;
-
-internal interface IInstructionMatcher
-{
-    bool Match(Instruction instruction, ValueMatchInterpolator outputs, InstructionPattern pattern);
-}
