+++ tests/DistIL.Tests/IR/MatchingTests.cs
@@ -9,15 +9,14 @@ public class MatchingTests
     {
         var inst = new BinaryInst(BinaryOp.Add, ConstInt.CreateI(42), new BinaryInst(BinaryOp.Mul, ConstInt.CreateI(1), ConstInt.CreateI(3)));
 
-        Instruction? instr = null;
-        if (inst.Match($"(add 42 {instr})")) {
-            Console.WriteLine("Right: " + instr);
-        }
+        BinaryInst? instr = null;
+        Assert.True(inst.Match($"(add 42 {instr})"));
+        Assert.IsType<BinaryInst>(instr);
+        Assert.Equal(BinaryOp.Mul, instr.Op);
 
         ConstInt? x = null;
-        if (inst.Match($"(add {x} (mul ? ?))"))
-        {
-            Console.WriteLine($"x: {x.Value}");
-        }
+        Assert.True(inst.Match($"(add {x} (mul ? ?))"));
+        Assert.IsType<ConstInt>(x);
+        Assert.Equal(42L, x.Value);
     }
 }
